% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woah_color_palette.R
\docType{data}
\name{woah_palette}
\alias{woah_palette}
\title{WOAH Corporate Color Palette}
\format{
A named character vector with 50 HEX color codes.
}
\usage{
woah_palette
}
\description{
A named vector of HEX codes used by the World Organisation for Animal Health (WOAH)
for consistent data visualization across reports, dashboards, and presentations.
}
\examples{
woah_palette["Blue"]
}
\keyword{datasets}
